chore: move root reducer and middlewares
