+++ src/modules/common/store/middlewares/loadingBar.js
