+++ src/modules/common/store/middlewares/loadingBar.test.js
