+++ src/modules/wallet/store/middlewares/hwManager.js
