+++ src/modules/wallet/store/middlewares/middleware.js
@@ -10,7 +10,7 @@ import { getTransactions } from '@transaction/api';
 import i18n from 'src/utils/i18n/i18n';
 import blockActionTypes from '@block/store/actionTypes';
 import settingsActionTypes from 'src/modules/settings/store/actionTypes';
-import actionTypes from './actionTypes';
+import actionTypes from '../actionTypes';
 
 const filterIncomingTransactions = (transactions, account) =>
   transactions.filter(transaction => (
