+++ src/modules/wallet/store/middlewares/middleware.test.js
@@ -7,7 +7,7 @@ import blockActionTypes from '@block/store/actionTypes';
 import settingsActionTypes from 'src/modules/settings/store/actionTypes';
 import * as transactionApi from '@transaction/api';
 import { getAutoLogInData } from 'src/utils/login';
-import walletActionTypes from './actionTypes';
+import walletActionTypes from '../actionTypes';
 import middleware from './middleware';
 
 jest.mock('src/utils/history');
