+++ src/redux/middlewares.js
@@ -1,6 +1,6 @@
 import thunk from 'redux-thunk';
 
-import account from '@wallet/store/middleware';
+import account from '@wallet/store/middlewares/middleware';
 import auth from '@auth/store/middleware';
 import block from '@block/store/middleware';
 import settings from 'src/modules/settings/store/middleware';
@@ -8,8 +8,8 @@ import bookmarks from '@bookmark/store/middleware';
 import network from '@network/store/middleware';
 import watchList from '@dpos/validator/store/middlewares/watchList';
 import voting from '@dpos/validator/store/middlewares/voting';
-import hwManager from './hwManager';
-import loading from './loadingBar';
+import hwManager from '@wallet/store/middlewares/hwManager';
+import loading from 'src/modules/common/store/middlewares/loadingBar';
 
 export default [
   account,
