+++ tests/__mocks__/redux-persist/index.js
@@ -1,6 +1,6 @@
 const real = jest.requireActual('redux-persist');
 module.exports = {
   ...real,
-  persistReducer: jest.fn().mockImplementation((config, reducers) => reducers),
-  persistStore: jest.fn().mockImplementation((store) => store),
+  persistReducer: jest.fn((config, reducers) => reducers),
+  persistStore: jest.fn((store) => store),
 };
